added create view edit journal entry added view reminder entry
