+++ amide_app/lib/components/reminder-tile.dart
@@ -1,4 +1,5 @@
-import 'package:amide_app/pages/edit/create-reminder-page.dart';
+import 'package:amide_app/pages/edit/edit-reminder-page.dart';
+import 'package:amide_app/pages/view/view-reminder-page.dart';
 import 'package:amide_app/utils/colors.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_slidable/flutter_slidable.dart';
@@ -65,65 +66,75 @@ class _ReminderTileState extends State<ReminderTile> {
         ),
         child: Padding(
           padding: const EdgeInsets.symmetric(horizontal: 20),
-          child: Container(
-            height: 80,
-            decoration: BoxDecoration(
-              color: AppColors.primBlue,
-              borderRadius: BorderRadius.circular(15),
-            ),
-            child: Padding(
-              padding: const EdgeInsets.all(20.0),
-              child: Row(
-                crossAxisAlignment: CrossAxisAlignment.center,
-                children: [
-                  Icon(
-                    Icons.event_note,
-                    color: value == true? Colors.white : Colors.grey[500],
-                    size: 25,
+          child: GestureDetector(
+            onTap: () {
+              Navigator.of(context).push(
+                  PageTransition(
+                    child: ViewReminder(),
+                    type: PageTransitionType.rightToLeft,
                   ),
-                  SizedBox(width: 30),
-                  Expanded(
-                    child: Column(
-                      crossAxisAlignment: CrossAxisAlignment.center,
-                      mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                      children: [
-                        Text(
-                          '08:00 AM',
-                          overflow: TextOverflow.ellipsis,
-                          style: TextStyle(
-                            decoration: TextDecoration.none,
-                            fontWeight: FontWeight.w600,
-                            fontSize: 16,
-                            fontFamily: 'Montserrat',
-                            color: value == true? Colors.white : Colors.grey[500],
-                          ),
-                        ),
-                        Text(
-                          'Breakfast',
-                          overflow: TextOverflow.ellipsis,
-                          style: TextStyle(
-                            fontWeight: FontWeight.w500,
-                            fontSize: 12,
-                            fontFamily: 'Montserrat',
-                            color: value == true? Colors.white : Colors.grey[500],
-                          ),
-                        )
-                      ],
+                );
+            },
+            child: Container(
+              height: 80,
+              decoration: BoxDecoration(
+                color: AppColors.primBlue,
+                borderRadius: BorderRadius.circular(15),
+              ),
+              child: Padding(
+                padding: const EdgeInsets.all(20.0),
+                child: Row(
+                  crossAxisAlignment: CrossAxisAlignment.center,
+                  children: [
+                    Icon(
+                      Icons.event_note,
+                      color: value == true? Colors.white : Colors.grey[500],
+                      size: 25,
                     ),
-                  ),
-                  SizedBox(width: 15),
-                  Switch.adaptive(
-                    activeColor: Colors.white,
-                    activeTrackColor: Color.fromARGB(255, 20, 44, 223),
-                    inactiveThumbColor: Colors.grey[500],
-                    inactiveTrackColor: Color.fromARGB(255, 10, 23, 119),
-                    value: value, 
-                    onChanged: (value) => setState(() {
-                      this.value = value;
-                      print(value);
-                    })
-                  )
-                ]
+                    SizedBox(width: 30),
+                    Expanded(
+                      child: Column(
+                        crossAxisAlignment: CrossAxisAlignment.center,
+                        mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                        children: [
+                          Text(
+                            '08:00 AM',
+                            overflow: TextOverflow.ellipsis,
+                            style: TextStyle(
+                              decoration: TextDecoration.none,
+                              fontWeight: FontWeight.w600,
+                              fontSize: 16,
+                              fontFamily: 'Montserrat',
+                              color: value == true? Colors.white : Colors.grey[500],
+                            ),
+                          ),
+                          Text(
+                            'Breakfast',
+                            overflow: TextOverflow.ellipsis,
+                            style: TextStyle(
+                              fontWeight: FontWeight.w500,
+                              fontSize: 12,
+                              fontFamily: 'Montserrat',
+                              color: value == true? Colors.white : Colors.grey[500],
+                            ),
+                          )
+                        ],
+                      ),
+                    ),
+                    SizedBox(width: 15),
+                    Switch.adaptive(
+                      activeColor: Colors.white,
+                      activeTrackColor: Color.fromARGB(255, 20, 44, 223),
+                      inactiveThumbColor: Colors.grey[500],
+                      inactiveTrackColor: Color.fromARGB(255, 10, 23, 119),
+                      value: value, 
+                      onChanged: (value) => setState(() {
+                        this.value = value;
+                        print(value);
+                      })
+                    )
+                  ]
+                ),
               ),
             ),
           ),
