+++ amide_app/lib/main.dart
@@ -1,7 +1,10 @@
+import 'package:amide_app/pages/create/create-journal-page.dart';
 import 'package:amide_app/pages/create/create-reminder-page.dart';
 import 'package:amide_app/pages/dashboard/dashboard-page.dart';
 import 'package:amide_app/pages/journal/journal-page.dart';
 import 'package:amide_app/pages/reminder/reminder-page.dart';
+import 'package:amide_app/pages/view/view-journal-page.dart';
+import 'package:amide_app/pages/view/view-reminder-page.dart';
 import 'package:flutter/material.dart';
 
 void main() {
@@ -30,7 +33,7 @@ class MyApp extends StatelessWidget {
         primarySwatch: Colors.blueGrey,
         fontFamily: 'Montserrat'
       ),
-      home: CreateReminder(),
+      home: DashboardPage(),
     );
   }
 }
